requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
langchain==0.3.9
langchain-openai==0.2.6
python-dotenv==1.0.1

motor==3.5.1
PyJWT==2.9.0
bcrypt==4.3.0
httpx==0.27.2
pymongo==4.6.3
dnspython>=2.3.0
email-validator>=2.0.0

razorpay==1.4.2
python-multipart>=0.0.9